Implement @preserveInvocation annotation
This commit implements the following subset of the slick.lifted api:

  - select *
  - take
  - map on single columns (no tuples, yet)
  - flatMap on single columns
  - filter for boolean literal conditions and < and > on int columns

The @preserveInvocation reifies the tree in the same order as it was
invoked. Example:

directQuery.take(1) => lift(directQuery).take(lift(1))

The benefit with with approach is that we fully leverage the existing
slick.lifted api. We only have to implement a few missing pieces that
don't align perfectly with slick.lifted (e.g., ==, $plus).

The current problem with this approach is that the reference to the
original TableQuery is lost. Compare a direct query column:

  DumpInfo(Path,@937946324.name,: String,Vector())

with a lifted query column:

  DumpInfo(Select,name,: String',Vector((in,Table User)))

Because of this, the value of two concatenated column strings is wrong.